Use F9 for refreshing buffer while preserving mode
Useful for picking up compilation output in cases where compilation
not easily can be made from within Emacs. Then we can do this in a
shell:

...build command... | tee compiler_output.txt

First time we open the compiler_output.txt and set it to
compilation-mode. Then each time we run the build command we only need
to press F9 to refresh.

Change-Id: I5eee09ab9cc43cf401ccfc16eef03f3cf1a81a93